rename folder to components
