+++ app/(app)/alpha/additional-details/components/divider.tsx
