+++ app/(app)/alpha/additional-details/components/text.tsx
