--- app/(app)/alpha/additional-details/tailwind-catalyts/index.ts
@@ -1,14 +1,0 @@
-export {
-  ErrorMessage,
-  Field,
-  Label,
-  Fieldset,
-  Legend,
-  Description,
-  FieldGroup,
-} from "./fieldset";
-export { Heading, Subheading } from "./heading";
-export { Divider } from "./divider";
-export { Input, InputGroup } from "./input";
-export { Select } from "./select";
-export { Button, TouchTarget } from "./button";
